backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.3.0
redis[hiredis]>=5.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
xmltodict>=0.13.0
chardet>=5.2.0
apscheduler>=3.10.0
prometheus-fastapi-instrumentator>=7.0.0
